refactor: Banner moves to components
