+++ src/App.js
@@ -6,7 +6,7 @@ import "bootstrap/dist/css/bootstrap.min.css";
 
 // layouts
 import MainLayout from "./components/MainLayout";
-import ApplyBanner from "./components/applyBanner/ApplyBanner";
+import ApplyBanner from "./components/Banner/Banner";
 import ScrollToTop from "components/ScrollToTop";
 
 // pages
