+++ src/pages/ApplyPage/components/Guide/ApplyGuide.js
@@ -5,12 +5,12 @@ import "../../../IntroducePage/components/styles/Introduce.css";
 import ApplyGuide_1 from "./ApplyGuide_1";
 import ApplyGuide_2 from "./ApplyGuide_2";
 
+import Banner from "../../../../components/Banner/Banner";
+
 const ApplyGuide = () => {
   return (
     <>
-      <div className="introducebanner">
-        <div className="bannerTitle">지원안내</div>
-      </div>
+      <Banner title="지원안내" />"
       <ApplyGuide_1 />
       <ApplyGuide_2 />
     </>
